add more instructions in comments
Because the readme is not in my working file structure for this script, it's helpful for me to have the each-time instructions in the script file itself. May help others as well.

When I learned that the bulk uploader would take a mix of domains, as long as they were valid, I began putting all of the RSS Feed Puller lists into a single file for uploading. But that still left the dozens of crawl outputs to upload individually.

I'm adding two post-processing steps - 
1. aggregate all the files that were created by the script into one file
2. split that aggregated file into the right sized chunks for uploading.

Taking these steps mean that we'll have only three or four files to upload, rather than dozens.